Use publicnode.com instead of ankr for public mainnet rpc
During the multicall stage of data sync, the ankr rpc recently started intermittently returning:
-32000 call retuned [sic] result on length exceeding limit 100000

Seems to be due to an update on the rpc side. Trying cloudflare and llamanode rpcs had the same result.
Publicnodes is working for now, but we should consider moving datasync to the backend and use
premium rpcs instead.